Added imgui docking

+++ FraplesSeven/src/Platform/WindowsInput.cpp
@@ -12,29 +12,26 @@ namespace Fraples
 	bool WindowsInput::IsKeyPressedImpl(int keyCoede)
 	{
 		auto window = static_cast<GLFWwindow*>(Application::GetApp().GetWindow().GetNativeWindow());
-
 		auto state = glfwGetKey(window, keyCoede);
-
 		return state == GLFW_PRESS || state == GLFW_REPEAT;
 	}
 	bool WindowsInput::IsMouseButtonPressedImpl(int button)
 	{
 		auto window = static_cast<GLFWwindow*>(Application::GetApp().GetWindow().GetNativeWindow());
 		auto state = glfwGetMouseButton(window,button);
-
 		return state == GLFW_PRESS ;
 	}
 	float WindowsInput::GetMouseXImpl()
 	{
 		auto[x, y] = GetMousePosImpl();
 
-		return (float)x;
+		return x;
 	}
 	float WindowsInput::GetMouseYImpl()
 	{
 		auto[x, y] = GetMousePosImpl();
 
-		return (float)y;
+		return y;
 		
 	}
